:art: car-names 관련 함수들 모아서 모듈 생성
