+++ src/library/car-names.js
@@ -1,0 +1,23 @@
+export function isCorrectCarNames(carName) {
+  if (carName.length > 0 && carName.length <= 5) {
+    return true;
+  }
+
+  return false;
+}
+
+export function addCorrectCarNames(carNames) {
+  const correctCarNames = carNames.map(function (carName) {
+    carName = carName.trim();
+    if (!isCorrectCarNames(carName)) {
+      alert("모든 자동차의 이름을 5자 이하로 입력해주세요.");
+      inputCarNames.value = "";
+
+      return;
+    }
+
+    return carName;
+  });
+
+  return correctCarNames;
+}
